dt-bindings: media: sunxi-ir: add A31 compatible
Allwinner A31 have introduce a new new memory mapping and a
reset line. Which make a gap and require a specific bindings.

The difference in memory mapping are :

- In the configure register there is a new sample bit
  and Allwinner has introduced the active threshold feature.

- In the status register a new STAT bit is present.

Note: CGPO and DRQ_EN bits are removed on A31 but present on A13
and on new SoCs like A64/H6.
This is actually not an issue as these bits are togglable and new
SoCs have a dedicated bindings.

Introduce this bindings to make a difference since this generation
and declare the reset line required since A31.

Signed-off-by: Clément Péron <peron.clem@gmail.com>